+++ app/controllers/v1/items_controller.rb
@@ -6,7 +6,8 @@ module V1
     end
 
     def show
-      @item = Item.find(params[:id])
+      @item = Item.find_by_id(params[:id])
+      return render json: not_found if @item.nil?
       render json: @item
     end
 
@@ -20,13 +21,15 @@ module V1
     end
 
     def update
-      @item = Item.find(params[:id])
+      @item = Item.find_by_id(params[:id])
+      return render json: not_found if @item.nil?
       @item.update(item_params) if item_params
       render json: @item
     end
 
     def destroy
-      @item = Item.find(params[:id])
+      @item = Item.find_by_id(params[:id])
+      return render json: not_found if @item.nil?
       if @item.destroy
         render json: { Deleted: "Item has been deleted" }
       end
@@ -36,6 +39,10 @@ module V1
 
     def item_params
       params.permit(:bucketlist_id, :name, :details, :done)
+    end
+
+    def not_found
+      { not_found!: "Item with id #{params[:id]} does not exist" }
     end
   end
 end
